start with a simple spawn
