+++ run.ts
@@ -1,5 +1,5 @@
-import { Reducer } from "./mod.ts";
 import type { Operation, Task } from "./types.ts";
+import { Reducer } from "./reduce.ts";
 
 export function run<T>(op: () => Operation<T>): Task<T> {
   let reducer = new Reducer();
