+++ spawn.ts
@@ -4,6 +4,6 @@ export function spawn<T>(block: () => Operation<T>): Operation<Task<T>> {
   return {
     *[Symbol.iterator]() {
       return (yield { type: "spawn", block }) as Task<T>;
-    }
-  }
+    },
+  };
 }
